Addressing initial PR feedback
- Removing MFEM exceptions
- Use long long int for true dof count
- SolveEstimateMarkRefine returns void
- Move adaptation configuration options to Refinement level
- Make dimensionalization of the mesh in PostOperator consistent

+++ palace/drivers/drivensolver.hpp
@@ -60,7 +60,7 @@ private:
                               const WavePortOperator &wave_port_op, int step,
                               double omega) const;
 
-  std::pair<ErrorIndicator, int>
+  std::pair<ErrorIndicator, long long int>
   Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const override;
 
 public:
