+++ palace/drivers/magnetostaticsolver.cpp
@@ -18,7 +18,7 @@
 namespace palace
 {
 
-std::pair<ErrorIndicator, int>
+std::pair<ErrorIndicator, long long int>
 MagnetostaticSolver::Solve(const std::vector<std::unique_ptr<mfem::ParMesh>> &mesh) const
 {
   // Construct the system matrix defining the linear operator. Dirichlet boundaries are
